draft/backend/requirements.txt
fastapi==0.103.0
sqlalchemy==2.1.0
passlib[bcrypt]==1.7.4
uvicorn==0.23.0
python-jose==3.3.0
pydantic==2.1.1
